fix: resolve platform-specific clipboard compilation errors
- mingwX64: Fix type mismatch by converting GMEM_MOVEABLE to UInt
- tvOS: Replace unsupported UIPasteboard with no-op implementations
  (tvOS does not support clipboard/pasteboard operations)

+++ cmp-library/src/mingwMain/kotlin/com/mobilebytelabs/kmptoolkit/clipboard/Clipboard.mingw.kt
@@ -40,7 +40,7 @@ actual fun copyToClipboard(text: String): Boolean {
 
         val bytes = text.encodeToByteArray()
         val size = (bytes.size + 1).toULong()
-        val hMem = GlobalAlloc(GMEM_MOVEABLE, size)
+        val hMem = GlobalAlloc(GMEM_MOVEABLE.toUInt(), size)
 
         if (hMem == null) {
             CloseClipboard()
